Fixed tests

+++ lib/toolbox/value_index.h
@@ -11,7 +11,7 @@ extern "C" {
 /** Get the index of a int32_t array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns -1.
+ * If no suitable elements were found, the function returns values_count.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -19,12 +19,12 @@ extern "C" {
  *
  * @return value's index.
  */
-int32_t value_index_int32(const int32_t value, const int32_t values[], size_t values_count);
+size_t value_index_int32(const int32_t value, const int32_t values[], size_t values_count);
 
 /** Get the index of a uint32_t array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns -1.
+ * If no suitable elements were found, the function returns values_count.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -32,12 +32,12 @@ int32_t value_index_int32(const int32_t value, const int32_t values[], size_t va
  *
  * @return value's index.
  */
-int32_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t values_count);
+size_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t values_count);
 
 /** Get the index of a float array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns -1.
+ * If no suitable elements were found, the function returns values_count.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -45,12 +45,12 @@ int32_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t
  *
  * @return value's index.
  */
-int32_t value_index_float(const float value, const float values[], size_t values_count);
+size_t value_index_float(const float value, const float values[], size_t values_count);
 
 /** Get the index of a bool array element which is equal to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns -1.
+ * If no suitable elements were found, the function returns values_count.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -58,12 +58,12 @@ int32_t value_index_float(const float value, const float values[], size_t values
  *
  * @return value's index.
  */
-int32_t value_index_bool(const bool value, const bool values[], size_t values_count);
+size_t value_index_bool(const bool value, const bool values[], size_t values_count);
 
 /** Get the index of a string array element which is equal to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns -1.
+ * If no suitable elements were found, the function returns values_count.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -71,7 +71,7 @@ int32_t value_index_bool(const bool value, const bool values[], size_t values_co
  *
  * @return value's index.
  */
-int32_t value_index_string(const char* value, const char* const values[], size_t values_count);
+size_t value_index_string(const char* value, const char* const values[], size_t values_count);
 
 /**
  * @brief String-to-string mapping
@@ -84,6 +84,7 @@ int32_t value_index_string(const char* value, const char* const values[], size_t
  * @param   length Length of either array (the must have equal lengths)
  * @param   input Input string to map
  * 
+ * @return Mapped string or NULL if not found.
  */
 const char* value_index_map_string(
     const char* const src_array[],
